requirements.txt
requests
boto3
snowflake-connector-python
python-dotenv
pandas